pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "zima-blue-cli"
version = "0.1.0"
description = "Zima Blue CLI - Personal Agent Orchestration Platform"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "User", email = "user@example.com"}
]
keywords = ["agent", "ai", "cli", "automation", "kimi"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "typer>=0.12.0",
    "rich>=13.0.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "jinja2>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1.0",
]

[project.scripts]
zima = "zima.cli:app"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.coverage.run]
source = ["zima"]
omit = ["tests/*", "zima/core/daemon.py", "zima/core/scheduler.py", "zima/core/state_manager.py"]

[tool.coverage.report]
fail_under = 60
show_missing = true

[tool.hatch.build.targets.wheel]
packages = ["zima"]
